+++ tests/plugins/xcube/rules/test_any_spatial_data_var.py
@@ -1,8 +1,7 @@
 from xrlint.plugins.xcube.rules.any_spatial_data_var import AnySpatialDataVar
+from xrlint.testing import RuleTest, RuleTester
+
 from .test_grid_mapping_naming import make_dataset
-
-from xrlint.testing import RuleTester, RuleTest
-
 
 valid_dataset = make_dataset()
 invalid_dataset = valid_dataset.drop_vars(["chl", "tsm"])