+++ tests/plugins/xcube/rules/test_lat_lon_naming.py
@@ -1,10 +1,8 @@
 import numpy as np
-
-from xrlint.plugins.xcube.rules.lat_lon_naming import LatLonNaming
-
 import xarray as xr
 
-from xrlint.testing import RuleTester, RuleTest
+from xrlint.plugins.xcube.rules.lat_lon_naming import LatLonNaming
+from xrlint.testing import RuleTest, RuleTester
 
 
 def make_dataset(lat_dim: str, lon_dim: str):