+++ tests/test_plugin.py
@@ -3,8 +3,8 @@ from unittest import TestCase
 
 import xarray as xr
 
-from xrlint.plugin import new_plugin, Plugin, PluginMeta
-from xrlint.processor import ProcessorOp, Processor
+from xrlint.plugin import Plugin, PluginMeta, new_plugin
+from xrlint.processor import Processor, ProcessorOp
 from xrlint.result import Message
 from xrlint.rule import Rule, RuleOp, define_rule
 
@@ -59,10 +59,8 @@ class PluginMetaTest(TestCase):
 
 
 class PluginDefineRuleDecoratorTest(TestCase):
-
     # noinspection PyUnusedLocal
     def test_decorator(self):
-
         plugin = Plugin(meta=PluginMeta(name="test"))
 
         @plugin.define_rule("my-rule-1")
@@ -97,10 +95,8 @@ class PluginDefineRuleDecoratorTest(TestCase):
 
 
 class PluginDefineProcessorDecoratorTest(TestCase):
-
     # noinspection PyUnusedLocal
     def test_decorator(self):
-
         plugin = Plugin(meta=PluginMeta(name="test"))
 
         @plugin.define_processor("my-processor-1")