+++ xrlint/formatters/json.py
@@ -1,9 +1,9 @@
 import json
 from collections.abc import Iterable
 
-from xrlint.formatter import FormatterOp, FormatterContext
-from xrlint.result import Result, get_rules_meta_for_results
+from xrlint.formatter import FormatterContext, FormatterOp
 from xrlint.formatters import registry
+from xrlint.result import Result, get_rules_meta_for_results
 from xrlint.util.schema import schema
 
 
@@ -19,7 +19,6 @@ from xrlint.util.schema import schema
     ),
 )
 class Json(FormatterOp):
-
     def __init__(self, indent: int = 2, with_meta: bool = False):
         super().__init__()
         self.indent = indent