+++ xrlint/plugins/core/rules/grid_mappings.py
@@ -1,6 +1,6 @@
 from xrlint.node import DatasetNode
 from xrlint.plugins.core.rules import plugin
-from xrlint.rule import RuleOp, RuleContext
+from xrlint.rule import RuleContext, RuleOp
 
 
 @plugin.define_rule(